Add icon `brand-metabrainz`
I originally wanted to have an icon for [MusicBrainz] in #760, but the
small icon without details which worked best as a Tabler icon is also
used for [other projects] by the [MetaBrainz Foundation].
The small icon variants are only distinguished by their colors, which
are not applicable here for a monochrome icon.

While MusicBrainz is the oldest and most widely known of these projects,
it makes more sense to name this brand icon after the foundation.
In order to make the icon discoverable by a search for the names of the
other projects, I have added tags for the main projects.

[MetaBrainz Foundation]: https://metabrainz.org/
[MusicBrainz]: https://musicbrainz.org/
[other projects]: https://metabrainz.org/projects